# openapi.yaml
openapi: 3.1.0
info:
  title: Buscador de Jurisprudencia Colombiana
  description: |
    Búsqueda híbrida (vectorial semántica + léxica) sobre jurisprudencia
    colombiana y doctrina jurídica indexada en una base vectorial Qdrant.
    Permite también obtener el texto OCR completo (paginado) de una sentencia
    ya identificada por un resultado de búsqueda.
  version: 1.1.0
servers:
  - url: https://chatgptactionjuris.onrender.com
    description: Proxy del Action desplegado en Render
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Organo:
      type: string
      description: |
        Fuente jurídica. Si se omite, busca en todas. Lista cerrada — si tu
        consulta es general, no envíes este campo.
      enum:
        - "Sala Civil - Corte Suprema de Justicia"
        - "Sala Laboral - Corte Suprema de Justicia"
        - "Sala Penal - Corte Suprema de Justicia"
        - "Sala Tutelas - Corte Suprema de Justicia"
        - "Sala Civil - Tribunal Superior de Medellín"
        - "Sala Laboral - Tribunal Superior de Medellín"
        - "Sala Penal - Tribunal Superior de Medellín"
        - "Doctrina"
    SearchResult:
      type: object
      properties:
        rank:
          type: integer
        filename:
          type: string
        organo:
          type: string
        file_path:
          type: string
        score:
          type: number
          nullable: true
        text_excerpt:
          type: string
        download_url:
          type: string
          nullable: true
          description: |
            URL pública para descargar el PDF original de la sentencia.
            Incluí este link al usuario como markdown clickeable cuando cites
            el documento.
    SearchResponse:
      type: object
      properties:
        query:
          type: string
        organo_filter:
          type: string
          nullable: true
        total_results:
          type: integer
        note:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
    DocumentResponse:
      type: object
      properties:
        filename:
          type: string
        file_path:
          type: string
        organo:
          type: string
          nullable: true
        num_pages:
          type: integer
        total_chars:
          type: integer
          description: Cantidad total de caracteres del documento.
        offset:
          type: integer
          description: Offset desde el que se leyó.
        returned_chars:
          type: integer
          description: Cantidad de caracteres devueltos en este request.
        has_more:
          type: boolean
          description: True si quedan más caracteres por leer.
        next_offset:
          type: integer
          nullable: true
          description: Offset para pedir el siguiente trozo. null si no quedan más.
        text:
          type: string
          description: Trozo del texto OCR del documento.
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - bearerAuth: []
paths:
  /search:
    get:
      operationId: buscarSentencias
      summary: Busca jurisprudencia y doctrina por consulta en lenguaje natural.
      description: |
        Devuelve fragmentos relevantes (ranked) de la base de conocimiento
        jurídico. Cada resultado incluye un text_excerpt truncado y un
        file_path con el cual se puede pedir el texto completo a /document.
      x-openai-isConsequential: false
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
          description: |
            Consulta jurídica en lenguaje natural. Puede ser una pregunta, un
            concepto o palabras clave.
        - in: query
          name: organo
          required: false
          schema:
            $ref: '#/components/schemas/Organo'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 10
          description: Cantidad máxima de fragmentos a devolver.
      responses:
        '200':
          description: Resultados de la búsqueda.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Petición inválida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token inválido o ausente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Error contactando al buscador upstream.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /document:
    get:
      operationId: obtenerTextoCompleto
      summary: Devuelve el texto OCR de una sentencia, paginado por caracteres.
      description: |
        Devuelve un trozo del texto OCR de una sentencia. El texto puede ser
        muy largo; usar offset y max_chars para paginar. Si has_more es true,
        llamar de nuevo con offset = next_offset para seguir leyendo.
      x-openai-isConsequential: false
      parameters:
        - in: query
          name: file_path
          required: true
          schema:
            type: string
          description: |
            Ruta relativa del archivo, tal como aparece en file_path de un
            resultado de /search.
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Offset (en caracteres) desde el que empezar a leer.
        - in: query
          name: max_chars
          required: false
          schema:
            type: integer
            minimum: 1000
            maximum: 80000
            default: 60000
          description: Máximo de caracteres a devolver en este request.
      responses:
        '200':
          description: Trozo del texto del documento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Petición inválida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token inválido o ausente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Error contactando al buscador upstream.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'